requirements.txt
torch>=2.1.0
transformers>=4.51.0
datasets>=2.19.0
accelerate>=0.30.0
numpy>=1.24.0
scipy>=1.11.0
powerlaw>=1.5
matplotlib>=3.7.0
seaborn>=0.13.0
pyyaml>=6.0
tqdm>=4.65.0
pandas>=2.0.0
